Admin bugfixes

+++ admin/handlers/Delete.handler.php
@@ -8,13 +8,15 @@ if(!isset($_SESSION['ip']) || !isset($_SESSION['user']) || !isset($_SESSION['rol
 require_once('../Connection.php');
 require_once('../Config.php');
 
-$connection = new Connection();
-$connection->connect(Config::$DB_HOST, Config::$DB_USER, Config::$DB_PASSWORD);
-$connection->selectDatabase(Config::$DB);
+if($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $connection = new Connection();
+    $connection->connect(Config::$DB_HOST, Config::$DB_USER, Config::$DB_PASSWORD);
+    $connection->selectDatabase(Config::$DB);
+    
+    $building = $_POST['building'];
+    
+    mysql_query("DELETE FROM building WHERE buildingID='" . mysql_real_escape_string($building) . "'");
+}
 
-$building = $_POST['building'];
-
-mysql_query("DELETE FROM building WHERE buildingID='" . mysql_real_escape_string($building) . "'");
-
-header('Location: ../Home');
+header('Location: ../Home');    
 ?>
